Concurrently access user data example. * waitgroups * goroutines * channels
