+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/ashtishad/waytogo/anthonygg"
+
+func main() {
+	anthonygg.UserDataConcurrencyIntro()
+}
